+++ unusual-spending/src/main/java/io/github/jonarzz/kata/unusual/spending/payment/PaymentRegisteredEvent.java
@@ -2,7 +2,6 @@ package io.github.jonarzz.kata.unusual.spending.payment;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
-import java.math.BigInteger;
 import java.time.OffsetDateTime;
 import java.util.UUID;
 
@@ -11,10 +10,10 @@ record PaymentRegisteredEvent(
         UUID id,
         @NotNull(message = "Payer ID cannot be null")
         @Positive(message = "ID should be a positive integer")
-        BigInteger payerId,
+        Long payerId,
         @NotNull(message = "Payment details cannot be null")
         PaymentDetails details,
-        OffsetDateTime timestamp
+        OffsetDateTime timestamp // TODO move to details
 ) {
 
 }
